updated return value for plan objects

+++ lbplanner/services/plan/delete_deadline.php
@@ -19,6 +19,7 @@ namespace local_lbplanner_services;
 use external_api;
 use external_function_parameters;
 use external_single_structure;
+use external_multiple_structure;
 use external_value;
 use local_lbplanner\helpers\plan_helper;
 use local_lbplanner\helpers\user_helper;
@@ -66,21 +67,22 @@ class plan_delete_deadline extends external_api {
             throw new \moodle_exception('Access denied');
         }
 
+        if (!plan_helper::check_edit_permissions($planid, $userid)) {
+            throw new \Exception('Access denied');
+        }
+
         $DB->delete_records(
             plan_helper::DEADLINES_TABLE,
             array(
-                'userid' => $userid,
                 'planid' => $planid ,
                 'moduleid' => $moduleid
             )
         );
 
-        return array('message' => 'Sucessfull');
+        return plan_helper::get_plan($planid);
     }
 
     public static function delete_deadline_returns() {
-        return new external_single_structure(
-            array('message' => new external_value(PARAM_TEXT, 'Gives back if the clearing was succesfull'))
-        );
+        return plan_helper::plan_structure();
     }
 }
